Update build script, lint, fix tracing in examples
