+++ langgraph/langchain.config.js
@@ -1,0 +1,24 @@
+import { resolve, dirname } from "node:path";
+import { fileURLToPath } from "node:url";
+
+/**
+ * @param {string} relativePath
+ * @returns {string}
+ */
+function abs(relativePath) {
+  return resolve(dirname(fileURLToPath(import.meta.url)), relativePath);
+}
+
+
+export const config = {
+  internals: [/node\:/, /@langchain\/core\//],
+  entrypoints: {
+    index: "index",
+    pregel: "pregel/index",
+    prebuilt: "prebuilt/index",
+  },
+  tsConfigPath: resolve("./tsconfig.json"),
+  cjsSource: "./dist-cjs",
+  cjsDestination: "./dist",
+  abs,
+}
