--- langgraph/scripts/create-entrypoints.js
@@ -1,102 +1,0 @@
-import * as fs from "fs";
-import * as path from "path";
-
-// .gitignore
-const DEFAULT_GITIGNORE_PATHS = ["node_modules", "dist", ".yarn"];
-
-// This lists all the entrypoints for the library. Each key corresponds to an
-// importable path, eg. `import { AgentExecutor } from "langchain/agents"`.
-// The value is the path to the file in `src/` that exports the entrypoint.
-// This is used to generate the `exports` field in package.json.
-// Order is not important.
-const entrypoints = {
-  index: "index",
-  pregel: "pregel/index",
-  prebuilt: "prebuilt/index",
-};
-
-// Entrypoints in this list require an optional dependency to be installed.
-// Therefore they are not tested in the generated test-exports-* packages.
-const requiresOptionalDependency = [];
-
-const updateJsonFile = (relativePath, updateFunction) => {
-  const contents = fs.readFileSync(relativePath).toString();
-  const res = updateFunction(JSON.parse(contents));
-  fs.writeFileSync(relativePath, JSON.stringify(res, null, 2) + "\n");
-};
-
-const generateFiles = () => {
-  const files = [...Object.entries(entrypoints), ["index", "index"]].flatMap(
-    ([key, value]) => {
-      const nrOfDots = key.split("/").length - 1;
-      const relativePath = "../".repeat(nrOfDots) || "./";
-      const compiledPath = `${relativePath}dist/${value}.js`;
-      return [
-        [
-          `${key}.cjs`,
-          `module.exports = require('${relativePath}dist/${value}.cjs');`,
-        ],
-        [`${key}.js`, `export * from '${compiledPath}'`],
-        [`${key}.d.ts`, `export * from '${compiledPath}'`],
-      ];
-    }
-  );
-
-  return Object.fromEntries(files);
-};
-
-const updateConfig = () => {
-  const generatedFiles = generateFiles();
-  const filenames = Object.keys(generatedFiles);
-
-  // Update package.json `exports` and `files` fields
-  updateJsonFile("./package.json", (json) => ({
-    ...json,
-    exports: Object.assign(
-      Object.fromEntries(
-        [...Object.keys(entrypoints)].map((key) => {
-          let entryPoint = {
-            types: `./${key}.d.ts`,
-            import: `./${key}.js`,
-            require: `./${key}.cjs`,
-          };
-
-          return [key === "index" ? "." : `./${key}`, entryPoint];
-        })
-      ),
-      { "./package.json": "./package.json" }
-    ),
-    files: ["dist/", ...filenames],
-  }));
-
-  // Write generated files
-  Object.entries(generatedFiles).forEach(([filename, content]) => {
-    fs.mkdirSync(path.dirname(filename), { recursive: true });
-    fs.writeFileSync(filename, content);
-  });
-
-  // Update .gitignore
-  fs.writeFileSync(
-    "./.gitignore",
-    filenames.join("\n") + "\n" + DEFAULT_GITIGNORE_PATHS.join("\n") + "\n"
-  );
-};
-
-const cleanGenerated = () => {
-  const filenames = Object.keys(generateFiles());
-  filenames.forEach((fname) => {
-    try {
-      fs.unlinkSync(fname);
-    } catch {
-      // ignore error
-    }
-  });
-};
-
-const command = process.argv[2];
-
-if (command === "pre") {
-  cleanGenerated();
-} else {
-  updateConfig();
-}
